+++ src/Snippets/Extensions/Snippets.cs
@@ -8,30 +8,6 @@ namespace Snippets.Extensions;
 
 internal static class Snippets
 {
-    public static void ConfigureTelemetry()
-    {
-        #region configure-telemetry
-
-        var telemetryOptions = new TelemetryOptions
-        {
-            // Configure logging
-            LoggerFactory = LoggerFactory.Create(builder => builder.AddConsole())
-        };
-
-        // Configure enrichers
-        telemetryOptions.MeteringEnrichers.Add(new MyMeteringEnricher());
-
-        // Configure telemetry listeners
-        telemetryOptions.TelemetryListeners.Add(new MyTelemetryListener());
-
-        var builder = new ResiliencePipelineBuilder()
-            .AddTimeout(TimeSpan.FromSeconds(1))
-            .ConfigureTelemetry(telemetryOptions) // This method enables telemetry in the builder
-            .Build();
-
-        #endregion
-    }
-
     public static async Task AddResiliencePipeline()
     {
         #region add-resilience-pipeline
@@ -88,24 +64,4 @@ internal static class Snippets
     {
         public TimeSpan Timeout { get; set; }
     }
-
-    #region telemetry-listeners
-
-    internal class MyTelemetryListener : TelemetryListener
-    {
-        public override void Write<TResult, TArgs>(in TelemetryEventArguments<TResult, TArgs> args)
-        {
-            Console.WriteLine($"Telemetry event occurred: {args.Event.EventName}");
-        }
-    }
-
-    internal class MyMeteringEnricher : MeteringEnricher
-    {
-        public override void Enrich<TResult, TArgs>(in EnrichmentContext<TResult, TArgs> context)
-        {
-            context.Tags.Add(new("my-custom-tag", "custom-value"));
-        }
-    }
-
-    #endregion
 }
