+++ test/Polly.Extensions.Tests/Telemetry/TelemetryListenerImplTests.cs
@@ -30,7 +30,7 @@ public class TelemetryListenerImplTests : IDisposable
         source.AttemptDuration.Description.Should().Be("Tracks the duration of execution attempts.");
         source.AttemptDuration.Unit.Should().Be("ms");
 
-        source.ExecutionDuration.Description.Should().Be("The execution duration and execution results of resilience pipelines.");
+        source.ExecutionDuration.Description.Should().Be("The execution duration of resilience pipelines.");
         source.ExecutionDuration.Unit.Should().Be("ms");
     }
